\name{functions}
\alias{clusterVafs}
\alias{hardClusterAssignments}
\alias{clusterWithBmm}
\alias{bmm.filter.clusters}
\alias{initScClass}
\alias{drawScatterPlot}
\alias{addPoints}
\alias{addHighlightLegend}
\alias{getClusterColors}
\alias{getOneSampleVafs}
\alias{addCnToVafs}
\alias{excludeRegions}
\alias{cleanAndAddCN}
\alias{getPurity}
\alias{getDensity}
\alias{getPeaks}
\alias{reorderBetaClust}
\alias{reorderBinomialClust}
\alias{clusterWithBinomialBmm}
\alias{reorderGaussianClust}
\alias{clusterWithGaussianBmm}
\alias{calculate.self.overlap}
\alias{bmm.calculate.pvalue}
\alias{binomial.bmm.filter.clusters}
\alias{gaussian.bmm.filter.clusters}
\alias{compute.binomial.error.bars}
\alias{calculate.pvalue}
\alias{getConnectivityMatrix}
\alias{reorderClust}
\title{
  sciClone functions
}
\description{
  miscellaneous internal functions used by the sciClone package
}
