\name{functions}
\alias{clusterVafs}
\alias{hardClusterAssignments}
\alias{clusterWithBmm}
\alias{bmm.filter.clusters}
\alias{initScClass}
\alias{drawScatterPlot}
\alias{addPoints}
\alias{addHighlightLegend}
\alias{plot2dWithMargins}
\alias{getClusterColors}
\alias{getOneSampleVafs}
\alias{addCnToVafs}
\alias{excludeRegions}
\alias{cleanAndAddCN}
\alias{getPurity}
\alias{getDensity}
\alias{getPeaks}

\title{
  cloneCaller Functions
}
\description{
  miscellaneous internal functions used by the sciClone package
}
